+++ neural_lam/weather_dataset.py
@@ -597,10 +597,10 @@ class WeatherDataset(torch.utils.data.Dataset):
 
         da_datastore_state = getattr(self, f"da_{category}")
         da_grid_index = da_datastore_state.grid_index
-        da_state_feature = da_datastore_state.state_feature
+        da_feature = da_datastore_state[f"{category}_feature"]
 
         coords = {
-            f"{category}_feature": da_state_feature,
+            f"{category}_feature": da_feature,
             "grid_index": da_grid_index,
         }
         if add_time_as_dim: